Modified program such that all settings stored in MainConfig. MainConfig no longer tracked on Git. DefaultMainConfig is tracked instead, systems falls back to DefaultMainConfig if some entries miss in MainConfig or MainConfig miss. Bug with empty spaces in TaskList manifested again - I fixed it

+++ StartSampling.py
@@ -1,6 +1,9 @@
+import JSONoperators
 from TaskManagement import DoTask
 import signal
 import os
+import SystemCheck
+import JSONoperators
 
 def signal_handler(signal, frame):
         global interrupted
@@ -15,20 +18,26 @@ def Sampling():
         interrupted = False
         signal.signal(signal.SIGINT, signal_handler)
 
+        JSONoperators.MergeJSONConfigs("MainConfig","DefaultMainConfig")
 
 
-        handle = open(".VSCINUSE", 'w')
-        handle.close()
+        #failures_found = SystemCheck.SystemCheck("MainConfig")
+        failures_found = False
+
+        if not failures_found:
+
+                handle = open(".VSCINUSE", 'w')
+                handle.close()
 
 
-        while True:
-                DoTask()
+                while True:
+                        DoTask()
 
 
-                if interrupted:
-                        print(f"Sampling process terminated")
-                        os.system("rm .VSCINUSE")
-                        break
+                        if interrupted:
+                                print(f"Sampling process terminated")
+                                os.system("rm .VSCINUSE")
+                                break
